+++ include/headers/tcp_header.hpp
@@ -20,7 +20,7 @@ public:
 
 	std::string get_port(bool src);
 
-	void get_bitstring(std::vector<int8_t> &to_fill, int8_t fill_with) override;
+	void get_bitstring(std::vector<float> &to_fill, int8_t fill_with) override;
 
 	void get_bitstring_header(std::vector<std::string> &to_fill) override;
