+++ include/headers/udp_header.hpp
@@ -15,7 +15,7 @@ class UDPHeader : public PacketHeader {
     void print_header(FILE *out) override;
     uint32_t get_header_len() override;
     std::string get_port(bool src);
-    void get_bitstring(std::vector<int8_t> &to_fill, int8_t fill_with) override;
+    void get_bitstring(std::vector<float> &to_fill, int8_t fill_with) override;
     void get_bitstring_header(std::vector<std::string> &to_fill) override;
     /* Header Specific */
   private:
